apps/etl/requirements.txt
flake8==7.0.0
python-dotenv==1.0.1
isort==5.13.2
pydantic-settings==2.3.3
psycopg2-binary==2.9.9